requirements.txt
fastapi
Pillow
pydantic
torch
torchvision
uvicorn
gunicorn
